% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/autos-data.R
\docType{data}
\name{autos}
\alias{autos}
\title{Automobile data set}
\format{\preformatted{'data.frame':	205 obs. of  17 variables:
 $ wheel.base       : Factor w/ 53 levels "100,4","101,2",..: 25 25 33 53 51 53 12 12 12 52 ...
 $ length           : Factor w/ 75 levels "141,1","144,6",..: 23 23 30 46 46 48 68 68 68 50 ...
 $ width            : Factor w/ 44 levels "60,3","61,8",..: 9 9 17 22 24 23 41 41 41 30 ...
 $ height           : Factor w/ 49 levels "47,8","48,8",..: 2 2 14 26 26 19 37 37 38 13 ...
 $ curb.weight      : int  2548 2548 2823 2337 2824 2507 2844 2954 3086 3053 ...
 $ engine.type      : int  1 1 6 4 4 4 4 4 4 4 ...
 $ num.of.cylinders : int  4 4 6 4 5 5 5 5 5 5 ...
 $ engine.size      : int  130 130 152 109 136 136 136 136 131 131 ...
 $ fuel.system      : int  6 6 6 6 6 6 6 6 6 6 ...
 $ bore             : Factor w/ 39 levels "","2,54","2,68",..: 25 25 3 15 15 15 15 15 12 12 ...
 $ stroke           : Factor w/ 37 levels "","2,07","2,19",..: 6 6 29 26 26 26 26 26 26 26 ...
 $ compression.ratio: Factor w/ 32 levels "10","10,1","11,5",..: 24 24 24 1 16 20 20 20 18 11 ...
 $ horsepower       : int  111 111 154 102 115 110 110 110 140 160 ...
 $ peak.rpm         : int  5000 5000 5000 5500 5500 5500 5500 5500 5500 5500 ...
 $ city.mpg         : int  21 21 19 24 18 19 19 19 17 16 ...
 $ highway.mpg      : int  27 27 26 30 22 25 25 25 20 22 ...
 $ price            : int  13495 16500 16500 13950 17450 15250 17710 18920 23875 NA ...
}}
\source{
\href{https://archive.ics.uci.edu/ml/datasets/Automobile}
}
\usage{
data(autos)
}
\description{
Data from Machine Learning Repository, UCI. This data set just containg a
numerical columns of original data set obtained from the repository. The last
column (price) is the dependent variable, and the first sixteen columns are the
independents variables. This data set, is useful for regression problems.
}
\details{
Attribute Information:
    Attribute:                Attribute Range:
------------------    --------------------------------------
1. wheel-base:          continuous from 86,6 120,9
2. length:              continuous from 141,1 to 208,1
3. width:               continuous from 60,3 to 72,3
4. height:              continuous from 47,8 to 59,8
5. curb-weight:         continuous from 1488 to 4066
6. engine-type:         1 (dohc), 2 (dohcv), 3 (l), 4 (ohc),
                        5 (ohcf), 6 (ohcv), 7 (rotor)
7. num-of-cylinders:    8, 5, 4, 6, 3, 12, 2
8. engine-size:         continuous from 61 to 326
9. fuel-system:         1 (1bbl), 2 (2bbl), 3 (4bbl), 4 (idi),
                        5 (mfi), 6 (mpfi), 7 (spdi), 8 (spfi)
10. bore:               continuous from 2,54 to 3,94
11. stroke:             continuous from 2,07 to 4,17
12. compression-ratio:  continuous from 7 to 23
13. horsepower:         continuous from 48 to 288
14. peak-rpm:           continuous from 4150 to 6600
15. city-mpg:           continuous from 13 to 49
16. highway-mpg:        continuous from 16 to 54
17. price:              continuous from 5118 to 45400
}
\examples{
data(autos)
}
\keyword{datasets}

