% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NoiseDetection.R
\name{calculateDiff}
\alias{calculateDiff}
\title{Calculate Diff}
\usage{
calculateDiff(dataSet, inf, sup)
}
\arguments{
\item{dataSet}{an object of class data frame with a data set.}

\item{inf}{an integer that represent the first column that you want review.}

\item{sup}{an integer that represent the last column that you want review.}
}
\value{
an object of class data frame with a variation of all data set.
}
\description{
Function to calculate de variation of a dataSet.
}
\examples{
iris.x <- iris[,1:4]
diffValues <- calculateDiff(iris.x)
}
\seealso{
getColumnsNoise
}

