% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Plots.R
\name{DensityPlot}
\alias{DensityPlot}
\title{Density Plot (Plot)}
\usage{
DensityPlot(data, col)
}
\arguments{
\item{data}{an object of class data frame with the a data.}

\item{col}{an integer that specify the column that you want for make the plot.}
}
\description{
Generate a density plot for a specific column of the data.
}
\examples{
iris.x <- iris[,1:4]
outlier.scores <- lof(iris.x, k = c(5:10))
DensityPlot(outlier.scores, 1)
}

