% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RegressionLibs.R
\name{diagnosticData}
\alias{diagnosticData}
\title{Diagnostic Data Generator}
\usage{
diagnosticData(fit)
}
\arguments{
\item{fit}{an object of class lm that contains the result to fit a linear model
regression.}
}
\value{
a data frame object with origin values, fitted values, and differents
error types for make some diagnostics plots.
}
\description{
Calculate different error types on the given data and return an object of class
data frame ready for usage in different diagnostics plots.
}
\examples{
iris.x <- iris[,1:3]
Petal.Width <- iris[,4]
ir.pca <- prcomp(iris.x, center = TRUE, scale. = TRUE)
PCA <- as.data.frame(ir.pca$x)

PC1 <- PCA[,1]
PC2 <- PCA[,2]
PC3 <- PCA[,3]

fit <- lm(Petal.Width ~ PC1 + PC2 + PC3, data = PCA)
diagnostic <- diagnosticData(fit)
}
\seealso{
ResidualsFitted, StResidualsFitted, NormalQQ, StResidualsLeverange
}

