% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Plots.R
\name{elbowPlot}
\alias{elbowPlot}
\title{Elbow Plot for PCA (Plot)}
\usage{
elbowPlot(data.pca)
}
\arguments{
\item{data.pca}{a list with class "prcomp" containing all principal components
calculated.}
}
\description{
Generate a plot of 10 first variances of Principal Components. This is useful to
determinate which are the most important components.
}
\examples{
iris.x <- iris[,1:4]
ir.pca <- prcomp(iris.x, center = TRUE, scale. = TRUE)

elbowPlot(ir.pca)
}
\seealso{
CalculateVariance, plotPC
}

