% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RegressionLibs.R
\name{findMissingValues}
\alias{findMissingValues}
\title{Find Missing Values}
\usage{
findMissingValues(dataSet)
}
\arguments{
\item{dataSet}{an object of class data frame with a data set.}
}
\value{
an integer(0) when no null values, or a data frame with all null values
identified by their positions (i,j)
}
\description{
Check out all data set to find missing values. Return a data frame, with the
possition of each missing value.
}
\examples{
#Example 1
# Getting a data set with missing values
cars <- read.csv("https://dl.dropboxusercontent.com/u/12599702/autos.csv", sep = ";", dec = ",")

missingValues <- findMissingValues(cars) #Find missing values

if (any(missingValues)) { #Validation if missing values exist
    missingValues <- missingValues[!duplicated(missingValues[,1]),] #Deleting duplicated rows
    print(missingValues) # print all missing values found
    cars <- cars[-missingValues[,1],] # deleting all rows that contains a missing values
}
}

