% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NoiseDetection.R
\name{getColumnsNoise}
\alias{getColumnsNoise}
\title{Columns Noise}
\usage{
getColumnsNoise(data, limit, inf, sup)
}
\arguments{
\item{data}{an object of class data frame with a data set of variation.}

\item{limit}{an double that represent the umbral for detect columns variation.}

\item{inf}{an integer that represent the first column that you want review.}

\item{sup}{an integer that represent the last column that you want review.}
}
\value{
an object of class data frame all columns that contains noise.
}
\description{
Function to calculate de variation of a dataSet.
}
\examples{
iris.x <- iris[,1:4]
diffValues <- calculateDiff(iris.x)
limit = 0.15
columnsNoise <- getColumnsNoise(diffValues, limit)
}
\seealso{
getColumnsNoise
}

