% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RegressionLibs.R
\name{makePairs}
\alias{makePairs}
\title{Make Pairs Function}
\source{
https://gastonsanchez.wordpress.com/2012/08/27/scatterplot-matrices-with-ggplot/
}
\usage{
makePairs(dataSet)
}
\arguments{
\item{dataSet}{an object of class data frame with a data set.}
}
\value{
an object of class data frame with the data used for ggplot function.
}
\description{
Function that generate a data frame with the data used for ggplot function for
make a scatterplot matrix.
}
\examples{
#This function is used by ScatterplotMatrix
#Example 1
iris.x <- iris[,1:4] # These are the independent variables
Species <- iris[,5] # This is the dependent variable

# A Scatterplot of all columns
ScatterplotMatrix(iris.x, c(1,2,3,4), Species, "Species")
# A Scatterplot of somes columns and different point size and alpha point
ScatterplotMatrix(iris.x, c(2,4), Species, "Species", 2, 1)
}
\seealso{
ScatterplotMatrix
}

