% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Plots.R
\name{MatPlot}
\alias{MatPlot}
\title{Plot Columns of Matrices (Plot)}
\usage{
MatPlot(dataSet, dependentVariable, dependentVariableName, from, to, x_lab)
}
\arguments{
\item{dataSet}{an object of class data frame with the a data set (Independent
variables).}

\item{dependentVariable}{is a list of values containig the dependent variable
of your regression model.}

\item{dependentVariableName}{is an optional parameter. It's an string that
contains de name of your dependent variable of your regression model.}

\item{from}{an integer that represent the first instance that you want in the plot.}

\item{to}{an integer that represent the last instance that you want in the plot.}

\item{x_lab}{a boolean that represent if you want or not the x axis scale. In
some cases, when you have many columns the plot could be ugly! The default value
is False.}
}
\description{
Generate a plot of the columns of a data set for all or a range of instances. In
some cases this is useful to identify some patron.
}
\examples{
iris.x <- iris[,1:4]
dependetVariable <- iris[,5]
MatPlot(iris.x, dependetVariable, "Species")
MatPlot(iris.x, dependetVariable, "Species", x_lab = TRUE)
}

