% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RegressionLibs.R
\name{myColorRamp}
\alias{myColorRamp}
\title{Color Ramp}
\source{
http://stackoverflow.com/questions/10413678/how-to-assign-color-scale-to-a-variable-in-a-3d-scatter-plot
}
\usage{
myColorRamp(colors, values)
}
\arguments{
\item{colors}{a list of name colors.}

\item{values}{an object of class data frame with a dependent variable.}
}
\value{
a list colors in HEX format.
}
\description{
Function that transforms a list of values in their corresponding color in the
given list.
}
\examples{
iris.y <- iris[,4]
cols <- myColorRamp(c("darkred", "yellow", "darkgreen"), iris.y)
}
\seealso{
PlotPC3D
}

