% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RegressionLibs.R
\name{normalize}
\alias{normalize}
\title{Normalize Function}
\usage{
normalize(dataSet)
}
\arguments{
\item{dataSet}{an object of class data frame with the original data set.}
}
\value{
an object of class data frame with a modified data set normalized
in range of 0 to 1.
}
\description{
Function to normalize data set in range of 0 to 1.
}
\examples{
#Example 1
iris.x <- iris[,1:4] #Just numerical columns witout missing values
normedIris <- as.data.frame(lapply(iris.x, normalize)) # In range [0,1]


#Example 2
# Getting a data set without missing values
cars <- read.csv("https://dl.dropboxusercontent.com/u/12599702/autosclean.csv", sep = ";", dec = ",")

normedCars <- as.data.frame(lapply(cars, normalize)) # In range [0,1]
}
\seealso{
scaleData, normalizeData
}

