% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RegressionLibs.R
\name{normalizeData}
\alias{normalizeData}
\title{Normalize Data Set}
\usage{
normalizeData(dataSet, min, max)
}
\arguments{
\item{dataSet}{an object of class data frame with the original data set that you
want normalize.}

\item{min}{an integer with the min value that you want normalize the data set.}

\item{max}{an integer with the max value that you want normalize the data set.}
}
\value{
an object of class data frame with a modified data set normalized
in the defined range.
}
\description{
Function to normalize a data set. You could normalize in any range [min,max], but
if you don't specify this values, this function make a normalization in range
[0,1].
}
\examples{
#Example 1
iris.x <- iris[,1:4] #Just numerical columns witout missing values

normedIris1 <- normalizeData(iris.x) #In range [0,1]
normedIris2 <- normalizeData(iris.x, 1, 10) #In ragen [1,10]


#Example 2
# Getting a data set without missing values
cars <- read.csv("https://dl.dropboxusercontent.com/u/12599702/autosclean.csv", sep = ";", dec = ",")

normedCars1 <- normalizeData(cars) #In range [0,1]
normedCars2 <- normalizeData(cars, 1, 10) #In ragen [1,10]
}
\seealso{
normalize, scaleData
}

