% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DiagnosticsPlots.R
\name{NormalQQ}
\alias{NormalQQ}
\title{Normal-QQ (Plot)}
\usage{
NormalQQ(diagnostic, dependentVariableName)
}
\arguments{
\item{diagnostic}{an object of class data frame with differents error types
calculated for make the graph.}

\item{dependentVariableName}{is an optional parameter. It's an string that
contains de name of your dependent variable of your regression model.}
}
\description{
Generate a plot of theoretical quantiles v/s Standarized Residuals of a
regression model.
}
\note{
Part of this code, it's from
http://librestats.com/2012/06/11/autoplot-graphical-methods-with-ggplot2/
}
\examples{
iris.x <- iris[,1:3]
Petal.Width <- iris[,4]
ir.pca <- prcomp(iris.x, center = TRUE, scale. = TRUE)
PCA <- as.data.frame(ir.pca$x)

PC1 <- PCA[,1]
PC2 <- PCA[,2]
PC3 <- PCA[,3]

fit <- lm(Petal.Width ~ PC1 + PC2 + PC3, data = PCA)
diagnostic <- diagnosticData(fit)
NormalQQ(diagnostic, "Petal Width")
}
\seealso{
diagnosticData, ResidualsFitted, StResidualsFitted, StResidualsLeverange
}

