% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Plots.R
\name{plotPC}
\alias{plotPC}
\title{Plot PCA (Plot)}
\usage{
plotPC(data.pca, dependentVariable, x_axis, y_axis, dependentVariableName)
}
\arguments{
\item{data.pca}{a list with class "prcomp" containing all principal components
calculated.}

\item{dependentVariable}{is a list of values containig the dependent variable
of your regression model.}

\item{x_axis}{an integer that represent the number of the principal component
that you want in your x axis.}

\item{y_axis}{an integer that represent the number of the principal component
that you want in your y axis.}

\item{dependentVariableName}{is an optional parameter. It's an string that
contains de name of your dependent variable of your regression model.}
}
\description{
Generate a plot of 2 Principal Components using ggplot. You must indicate which
PC you want in the graph.
}
\examples{
iris.x <- iris[,1:3]
Petal.Width <- iris[,4]
ir.pca <- prcomp(iris.x, center = TRUE, scale. = TRUE)

plotPC(ir.pca, Petal.Width, 1, 2, "Petal Width")
}
\seealso{
linePlot
}

