% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Plots.R
\name{PlotPC3D}
\alias{PlotPC3D}
\title{Plot PC 3D (3DPlot)}
\usage{
PlotPC3D(data.pca, from, to, dependentVariable)
}
\arguments{
\item{data.pca}{a list with class "prcomp" containing all principal components
calculated.}

\item{from}{an integer that represent the first principal component that you
want in the plot.}

\item{to}{an integer that represent the last principal component that you
want in the plot.}

\item{dependentVariable}{is a list of values containig the dependent variable
of your regression model.}
}
\description{
Generate a 3D plot for a range of principal components.
}
\examples{
iris.x <- iris[,1:3]
Petal.Width <- iris[,4]
ir.pca <- prcomp(iris.x, center = TRUE, scale. = TRUE)
PlotPC3D(ir.pca, 1, 3, Petal.Width)
}

