% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RegressionLibs.R
\name{removeRowsMissing}
\alias{removeRowsMissing}
\title{Remove Missing Values of Data Set}
\usage{
removeRowsMissing(missingValues, dataSet)
}
\arguments{
\item{missingValues}{an object of class Data.Frame with the possition (i,j) of
each missing value.}

\item{dataSet}{an object of class Data.Frame with the original Data Set.}
}
\value{
an object of class Data.Frame with a modified Data Set without missing
values.
}
\description{
Delete all rows of data set that contains a missing value.
}
\examples{
NA_values <- findMissingValues(iris)

if (any(NA_values)) { #Validation if missing values exist
   iris <- removeRowsMissing(NA_values, iris) #Remove all rows with missing values
}
}
\seealso{
findMissingValues
}

