% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RegressionLibs.R
\name{scaleData}
\alias{scaleData}
\title{Scale Function}
\usage{
scaleData(dataSet, min, max)
}
\arguments{
\item{dataSet}{an object of class data frame with the original data set.}

\item{min}{an integer with the min value that you want scale the data set.}

\item{max}{an integer with the max value that you want scale the data set.}
}
\value{
an object of class data frame with a modified data set scaled
in the defined range.
}
\description{
Function to scale the data set. If you use normalize function first and then
scaleData function you could obtain a normalize dataset in a range [x,y]
}
\examples{
#Example 1
iris.x <- iris[,1:4] #Just numerical columns witout missing values

normed <- as.data.frame(lapply(iris.x, normalize))
normedIris <- as.data.frame(lapply(normed, scaleData, 1, 10)) #In range [1,10]


#Example 2
# Getting a data set without missing values
cars <- read.csv("https://dl.dropboxusercontent.com/u/12599702/autosclean.csv", sep = ";", dec = ",")

normed <- as.data.frame(lapply(cars, normalize))
normedCars <- as.data.frame(lapply(normed, scaleData, 1, 10)) #In range [1,10]
}
\seealso{
normalize, normalizeData
}

