% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Plots.R
\name{ScatterplotMatrix}
\alias{ScatterplotMatrix}
\title{Scatterplot Matrix (Plot)}
\source{
https://gastonsanchez.wordpress.com/2012/08/27/scatterplot-matrices-with-ggplot/
}
\usage{
ScatterplotMatrix(data.pca, from, to, dependentVariable, dependentVariableName)
}
\arguments{
\item{dependentVariable}{is a list of values containig the dependent variable
of your regression model.}

\item{dependentVariableName}{is an optional parameter. It's an string that
contains de name of your dependent variable.}

\item{data}{an object of class "data.frame" containing just numerical columns.}

\item{columns}{an object of class "numeric" containing the list of columns
that you want in your scatterplot.}
}
\description{
Generate a Scatterplot Matrix of some columns of data set using ggplot.
}
\examples{
iris.x <- iris[,1:3]
columns <- c(1,2,3)
Petal.Width <- iris[,4]
ScatterplotMatrix(iris.x, columns, Petal.Width, "Petal Width")
}
\seealso{
makePairs
}

