% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RegressionLibs.R
\name{CalculateVariance}
\alias{CalculateVariance}
\title{Calculate Variance Function}
\usage{
CalculateVariance(dataSet, col)
}
\arguments{
\item{dataSet}{an object of class data frame with a data set that you want
calculate the variance.}

\item{col}{an integer that represent the column that you want calculate the
variance.}
}
\value{
an object of class data frame with a modified data set with variance
calculated.
}
\description{
Function to calculate de variance of an specific column.
}
\examples{
iris.x <- iris[,1:4]
ir.pca <- prcomp(iris.x, center = TRUE, scale. = TRUE)
rowsData <- length(ir.pca$sdev)
seqRow <- seq(from = 1, to = rowsData, length.out = rowsData)
dataPlot <- data.frame(seqRow, ir.pca$sdev)
dataPlot <- CalculateVariance(dataPlot, 2)
}
\seealso{
linePlot
}

