% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RegressionLibs.R
\name{findMissingValues}
\alias{findMissingValues}
\title{Find Missing Values}
\usage{
findMissingValues(dataSet)
}
\arguments{
\item{dataSet}{an object of class Data.Frame with a dataset.}
}
\value{
an integer(0) when no null values, or a Data.Frame with all null values
identified by their positions (i,j)
}
\description{
Check out all dataset to find missing values. Return a data frame, with the
possition of each missing value.
}
\examples{
NA_values <- findMissingValues(iris)

if (any(NA_values)) { #Validation if missing values exist
   #Do something with your missing values
}
}
\seealso{
removeRowsMissing
}

