% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RegressionLibs.R
\name{makePairs}
\alias{makePairs}
\title{Make Pairs Function}
\source{
https://gastonsanchez.wordpress.com/2012/08/27/scatterplot-matrices-with-ggplot/
}
\usage{
makePairs(dataSet)
}
\arguments{
\item{dataSet}{an object of class data frame with a data set.}
}
\value{
an object of class data frame with the data used for ggplot function.
}
\description{
Function that generate a data frame with the data used for ggplot function for
make a scatterplot matrix.
}
\examples{
iris.x <- iris[,1:4]
ir.pca <- prcomp(iris.x, center = TRUE, scale. = TRUE)
gg1 <- makePairs(ir.pca$x)
}
\seealso{
ScatterplotMatrix
}

