% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Plots.R
\name{PlotPC3D}
\alias{PlotPC3D}
\title{Plot PC 3D (3DPlot)}
\usage{
PlotPC3D(data, columns, dependentVariable)
}
\arguments{
\item{data}{an object of class data frame with the data.}

\item{columns}{an object of class "numeric" containing the list of columns
that you want in your parallel plot.}

\item{dependentVariable}{an object of class "numeric", "factor" or "integer" is
a list of values containig the dependent variable.}
}
\description{
Generate a 3D plot Generates a 3D graphic for a set of 3 columns of the data set.
}
\examples{
#Example 1
iris.x <- iris[,1:4] # These are the independent variables
Species <- iris[,5] # This is the dependent variable

# 3D Plot of 3 first columns of data set
PlotPC3D(iris.x, c(1,2,3), Species)


#Example 2
iris.x <- iris[,1:4] # These are the independent variables
Species <- iris[,5] # This is the dependent variable

ir.pca <- prcomp(iris.x, center = TRUE, scale. = TRUE) # performin prcomp

# 3D Plot of 3 first columns of data set
PlotPC3D(as.data.frame(ir.pca$x), c(1,2,3), Species)


#Example 3
# Getting a clean data set (without missing values)
cars <- read.csv("https://dl.dropboxusercontent.com/u/12599702/autosclean.csv", sep = ";", dec = ",")
cars.x <- cars[,1:16] # These are the independent variables
cars.y <- cars[,17] # This is the dependent variable

# 3D Plot of 3 first columns of data set
PlotPC3D(cars.x, c(1,2,3), cars.y)


#Example 4
# Getting a clean data set (without missing values)
cars <- read.csv("https://dl.dropboxusercontent.com/u/12599702/autosclean.csv", sep = ";", dec = ",")
cars.x <- cars[,1:16] # These are the independent variables
cars.y <- cars[,17] # This is the dependent variable

cars.pca <- prcomp(cars.x, center = TRUE, scale. = TRUE) # performin prcomp

# 3D Plot of 3 first columns of data set
PlotPC3D(as.data.frame(cars.pca$x), c(1,2,3), cars.y)
}

