% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RegressionLibs.R
\name{scaleData}
\alias{scaleData}
\title{Scale Function}
\usage{
scaleData(dataSet, min, max)
}
\arguments{
\item{dataSet}{an object of class data frame with the original data set.}

\item{min}{an integer with the min value that you want scale the data set.}

\item{max}{an integer with the max value that you want scale the data set.}
}
\value{
an object of class data frame with a modified data set scaled
in the defined range.
}
\description{
Function to scale the data set. If you use normalize function first and then
scaleData function you could obtain a normalize dataset in a range [x,y]
}
\examples{
iris.x <- iris[,1:4]
normed <- as.data.frame(lapply(iris.x, normalize))
normedIris <- as.data.frame(lapply(normed, 1, 10, scaleData)) #In range [1,10]
}
\seealso{
normalize, normalizeData
}

