% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Plots.R
\name{ScatterplotMatrix}
\alias{ScatterplotMatrix}
\title{Scatterplot Matrix (Plot)}
\source{
https://gastonsanchez.wordpress.com/2012/08/27/scatterplot-matrices-with-ggplot/
}
\usage{
ScatterplotMatrix(data, columns, dependentVariable, dependentVariableName,
  pointSize, alphaPoint)
}
\arguments{
\item{data}{an object of class "data.frame" containing just numerical columns.}

\item{columns}{an object of class "numeric" containing the list of columns
that you want in your scatterplot.}

\item{dependentVariable}{an object of class "numeric", "factor" or "integer" is
a list of values containig the dependent variable.}

\item{dependentVariableName}{is an optional parameter. It's an string that
contains de name of your dependent variable.}

\item{pointSize}{is an optional parameter of class numeric with a single value
that represent the point size of plot.}

\item{alphaPoint}{is an optional parameter of class numeric with a single value
that represent the alpha of points in the plot.}
}
\description{
Generate a Scatterplot Matrix of some columns of data set using ggplot.
}
\examples{
#Example 1
iris.x <- iris[,1:4] # These are the independent variables
Species <- iris[,5] # This is the dependent variable

# An Scatterplot of all columns
ScatterplotMatrix(iris.x, c(1,2,3,4), Species, "Species")
# An Scatterplot of somes columns and different point size and alpha point
ScatterplotMatrix(iris.x, c(2,4), Species, "Species", 2, 1)


#Example 2
}
\seealso{
makePairs
}

