% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Plots.R
\name{ScatterplotMatrix}
\alias{ScatterplotMatrix}
\title{Scatterplot Matrix (Plot)}
\source{
https://gastonsanchez.wordpress.com/2012/08/27/scatterplot-matrices-with-ggplot/
}
\usage{
ScatterplotMatrix(data.pca, from, to, dependentVariable, dependentVariableName)
}
\arguments{
\item{data.pca}{a list with class "prcomp" containing all principal components
calculated.}

\item{from}{an integer that represent the first principal component that you
want in the scatterplot matrix.}

\item{to}{an integer that represent the last principal component that you
want in the scatterplot matrix.}

\item{dependentVariable}{is a list of values containig the dependent variable
of your regression model.}

\item{dependentVariableName}{is an optional parameter. It's an string that
contains de name of your dependent variable of your regression model.}
}
\description{
Generate a Scatterplot Matrix between a range using ggplot.
}
\examples{
iris.x <- iris[,1:3]
Petal.Width <- iris[,4]
ir.pca <- prcomp(iris.x, center = TRUE, scale. = TRUE)

ScatterplotMatrix(ir.pca, 1, 3, Petal.Width, "Petal Width")
}
\seealso{
makePairs
}

